Redesign how ShareMemory works
The way ShareMemory works right now is the sender knows the recipient's
shared address and by sending the message to the service registry, the
appropriate recipient page tables are modified whether the recipient
accepts/knows or not. Change it so that the owner initiates with a
ShareMemoryRequest, ServiceRegistry sends a ShareMemoryInvitation
to the recipient, and then the recipient sends a ShareMemoryResponse
that either accepts with a share address or denies.

Implements #14.