Fix how ServiceRegistry handles ShareMemoryRequests
Memory sharing works at page granulity, so we need to tell the recipient
how many bytes worth of pages it needs to allocate, as well as the offset
from the beginning of the starting page to the start of the address the
sender wanted to share. Also sends the share result to both parties.