Explore GUI DataBinding
For the UI data description language, I'm thinking of allowing UI
element properties to be bound to some source. Eg for textboxes,

(textbox (bind text commandLine))

Where 'text' is the string value of the textbox, and 'commandLine' is
the name of a variable somewhere. Right now I'm thinking the
owner of 'commandLine' would define it as an Observable<char*>,
TextBox would then have a Bindable<char*> text property.
Writes (using setValue or probably assignment operator) to an
Observable notifies its subscriber (Binding), which notifies its owner
(TextBox), which then requests the value from the Observable.