Start designing Apollo's basic framework elements
UIElements have a desired width/height {bounds}, but its parent
container may give it a different actual bounds depending on
the container's config.

Elements can have MetaData associated with them, which function
similar to WPF's Attached Properties. In a layout file, you can add
a meta property to an element, which specify values that its parent
container use to arrange that element. For example:

(textbox
    (bind text commandLine)
    (meta (grid
        (row 1)
        (column 1))))

When the above textbox is nested in a Grid, the Grid will use
the specified row and column values to layout the textbox.