Add helper func to extract lists starting with symbols
For Apollo layout files, there are lists like (fixed-height 50). A list
that starts with a symbol (fixed-height in this case) is considered
to be a Constructor. A Constructor is an easy way to examine
a SExpression* to find the name of the constructor, the
values in the list, and a way to extract a derived type
from the list by index.

Part of #15