Window is now a Container
Window now ties into the Elements framework so that it can hold one
child container. This is because elements' getBounds needs to call its
parent to find out what its own bounds are, and Window already
keeps track of its own bounds which are the max bounds of the
whole tree, so it becomes the root.